Delayed::Job
Running only in development or production

+++ config/delayed_job.rb
@@ -34,7 +34,9 @@ Delayed::Worker.sleep_delay = 5
 Delayed::Worker.max_attempts = 5
 Delayed::Worker.max_run_time = 5.minutes
 
-Delayed::Job.delete_all
+Delayed::Job.delete_all if ENV.fetch('RAILS_ENV', nil) == 'test'
+
+# Load the database configuration
 
 config = YAML.load_file('config/database.yml')[ENV.fetch('RACK_ENV', 'development')]
 ActiveRecord::Base.establish_connection(config)
